requirements.txt
pandas==2.2.3
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.13.2
streamlit==1.40.2
streamlit_tags-1.2.8